settings performance optimization pass
