+++ applications/process-manager/qml.qrc
@@ -1,6 +1,6 @@
 <RCC>
     <qresource prefix="/">
         <file>main.qml</file>
-        <file alias="font/icomoon.ttf">../screenshot-viewer/font/icomoon.ttf</file>
+        <file alias="font/icomoon.ttf">font/icomoon.ttf</file>
     </qresource>
 </RCC>
